backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
openai>=1.3.0
azure-ai-documentintelligence>=1.0.0
azure-identity>=1.15.0
pydantic-settings>=2.0.0
websockets>=12.0
aiofiles>=23.2.1
